# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Listing {
  id: number;
  name: string;
  game: string;
  setName: string;
  rarity: string;
  condition: string;
  price: number;
  seller: string;
  image: string;
  listedAt: string;
  featured?: boolean;
  tradeable?: boolean;
}

export interface ListingInput {
  name: string;
  game: string;
  setName: string;
  rarity: string;
  condition: string;
  price: number;
  image: string;
  tradeable?: boolean;
}

export interface Pack {
  id: number;
  name: string;
  game: string;
  description: string;
  price: number;
  image: string;
  cardsPerPack: number;
  stock: number;
  accent?: string;
}

export interface MarketplaceSummary {
  activeListings: number;
  tradesToday: number;
  packsSold: number;
  averagePrice: number;
}

export interface Trade {
  id: number;
  offeredCard: string;
  requestedCard: string;
  trader: string;
  status: string;
  createdAt: string;
  image?: string;
}

export interface TradeInput {
  offeredCard: string;
  requestedCard: string;
  image?: string;
}

export interface Conversation {
  id: number;
  participant: string;
  listingName: string;
  preview: string;
  updatedAt: string;
  unread: number;
  avatar?: string;
}

export interface Message {
  id: number;
  conversationId: number;
  text: string;
  sender: string;
  sentAt: string;
  mine: boolean;
}

export interface MessageInput {
  /** @minLength 1 */
  text: string;
}

export type ListListingsParams = {
game?: string;
category?: string;
search?: string;
sort?: string;
};