# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  id: number;
  name: string;
  game: string;
  setName: string;
  rarity: string;
  condition: string;
  price: number;
  seller: string;
  image: string;
  listedAt: string;
  featured?: boolean;
  tradeable?: boolean;
}