# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ListingInput {
  name: string;
  game: string;
  setName: string;
  rarity: string;
  condition: string;
  price: number;
  image: string;
  tradeable?: boolean;
}