# lib/api-zod/src/generated/types/pack.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Pack {
  id: number;
  name: string;
  game: string;
  description: string;
  price: number;
  image: string;
  cardsPerPack: number;
  stock: number;
  accent?: string;
}