# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Trade {
  id: number;
  offeredCard: string;
  requestedCard: string;
  trader: string;
  status: string;
  createdAt: string;
  image?: string;
}